Onboarding Changes
- Changed from Activity to Fragment
- Added YoutubeView

TODO:
Add Choice of first Config Name
