+++ app/src/main/java/com/schneewittchen/rosandroid/ui/activity/MainActivity.java
@@ -1,8 +1,12 @@
 package com.schneewittchen.rosandroid.ui.activity;
 
 import android.Manifest;
+import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
+import android.view.Window;
+import android.view.WindowManager;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
@@ -10,6 +14,7 @@ import androidx.core.app.ActivityCompat;
 import androidx.fragment.app.Fragment;
 
 import com.schneewittchen.rosandroid.R;
+import com.schneewittchen.rosandroid.ui.fragments.IntroFragment;
 import com.schneewittchen.rosandroid.ui.fragments.MainFragment;
 import com.schneewittchen.rosandroid.ui.helper.OnBackPressedListener;
 
@@ -40,13 +45,19 @@ public class MainActivity extends AppCompatActivity {
 
         setContentView(R.layout.activity_main);
 
-        Toolbar myToolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(myToolbar);
-
-        if (savedInstanceState == null) {
+        if (savedInstanceState == null && !restorePrefData()) {
             getSupportFragmentManager().beginTransaction()
+                    .replace(R.id.main_container, IntroFragment.newInstance())
+                    .commitNow();
+        } else {
+            Toolbar myToolbar = findViewById(R.id.toolbar);
+            setSupportActionBar(myToolbar);
+
+            if (savedInstanceState == null) {
+                getSupportFragmentManager().beginTransaction()
                     .replace(R.id.main_container, MainFragment.newInstance())
                     .commitNow();
+            }
         }
 
         this.requestPermissions();
@@ -72,8 +83,12 @@ public class MainActivity extends AppCompatActivity {
         String[] permissions = new String[] {
                 Manifest.permission.ACCESS_NETWORK_STATE,
                 Manifest.permission.ACCESS_FINE_LOCATION};
-
         ActivityCompat.requestPermissions(this, permissions, LOCATION_PERM);
+    }
 
+    private boolean restorePrefData() {
+        SharedPreferences pref = getApplicationContext().getSharedPreferences("onboardingPrefs", MODE_PRIVATE);
+        Boolean checkedIn = pref.getBoolean("CheckedIn", false);
+        return checkedIn;
     }
 }
