+++ app/src/main/res/layout/layout_onboarding_screen.xml
@@ -6,16 +6,17 @@
     android:layout_height="match_parent">
 
     <TextView
-        android:textStyle="bold"
-        android:textSize="32dp"
         android:id="@+id/introTitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="TextView"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:textSize="32dp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toTopOf="@+id/introDescription"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/introImage" />
 
     <TextView
         android:id="@+id/introDescription"
@@ -28,8 +29,10 @@
         android:lineSpacingExtra="16dp"
         android:text="TextView"
         android:textAlignment="center"
+        android:textAppearance="@style/TextAppearance.AppCompat.Medium"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/introTitle"
         app:layout_constraintVertical_bias="0.13999999" />
@@ -41,6 +44,9 @@
         android:layout_marginBottom="32dp"
         app:layout_constraintBottom_toTopOf="@+id/introTitle"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_chainStyle="packed"
         tools:srcCompat="@tools:sample/avatars" />
 </androidx.constraintlayout.widget.ConstraintLayout>
